// src/types.ts
export interface NpcData {
  name: string;
  initBody: number;
  initDex: number;
  initMind: number;
  initSoul: number;
  meleePl: number;
  archeryPl: number;
  alchemyPl: number;
  magicPl: number;
  armor: number;
  extraHp: number;
  extraMeleeDmg: number;
  extraArcheryDmg: number;
  extraArcheryRange: number;
  extraIp: number;
  speeds: string;
}